[metal3] Rename health card to status card
